feature/level2 : 요구사항 5 - 미션 토글 시 점수 반영
